--- src/main/java/fr/craftinglabs/pi/apa102/LedStrip.java
@@ -1,5 +1,0 @@
-package fr.craftinglabs.pi.apa102;
-
-public interface LedStrip {
-    void setColors(Iterable<ARGBColor> colorBand);
-}
