+++ src/main/java/fr/craftinglabs/pi/apa102/matrix/Apa102LedStrip.java
@@ -1,4 +1,8 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.matrix;
+
+import fr.craftinglabs.pi.apa102.ARGBColor;
+import fr.craftinglabs.pi.apa102.matrix.io.LedStrip;
+import fr.craftinglabs.pi.apa102.matrix.io.LedStripIO;
 
 public class Apa102LedStrip implements LedStrip {
