+++ src/main/java/fr/craftinglabs/pi/apa102/matrix/ApaMatrix.java
@@ -1,4 +1,6 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.matrix;
+
+import fr.craftinglabs.pi.apa102.matrix.io.LedStrip;
 
 public class ApaMatrix {
