+++ src/main/java/fr/craftinglabs/pi/apa102/matrix/ColorBand.java
@@ -1,4 +1,6 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.matrix;
+
+import fr.craftinglabs.pi.apa102.ARGBColor;
 
 import java.util.ArrayList;
 import java.util.Collections;
