+++ src/main/java/fr/craftinglabs/pi/apa102/matrix/Frame.java
@@ -1,4 +1,6 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.matrix;
+
+import fr.craftinglabs.pi.apa102.ARGBColor;
 
 public class Frame {
     private ARGBColor frame[][];
