+++ src/main/java/fr/craftinglabs/pi/apa102/matrix/Size.java
@@ -1,4 +1,4 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.matrix;
 
 public class Size {
     private final int nbColumn;
