+++ src/main/java/fr/craftinglabs/pi/apa102/matrix/io/LedStripIO.java
@@ -1,4 +1,4 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.matrix.io;
 
 public interface LedStripIO {
     void shift(int[] bits);
