+++ src/main/java/fr/craftinglabs/pi/apa102/time/TimeFrameBuilder.java
@@ -1,5 +1,11 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.time;
 
+
+import fr.craftinglabs.pi.apa102.ARGBColor;
+import fr.craftinglabs.pi.apa102.matrix.Frame;
+import fr.craftinglabs.pi.apa102.matrix.Size;
+import fr.craftinglabs.pi.apa102.font.Font;
+import fr.craftinglabs.pi.apa102.font.Glyph;
 
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
