+++ src/test/java/fr/craftinglabs/pi/apa102/matrix/ApaMatrixTest.java
@@ -1,5 +1,7 @@
-package fr.craftinglabs.pi.apa102;
+package fr.craftinglabs.pi.apa102.matrix;
 
+import fr.craftinglabs.pi.apa102.ARGBColor;
+import fr.craftinglabs.pi.apa102.matrix.io.LedStrip;
 import org.junit.Before;
 import org.junit.Test;
